backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.3
python-dotenv==1.0.0
email-validator==2.1.0
Pillow==11.3.0
